move demo-related code to separate file
