+++ index.js
@@ -1,45 +1,9 @@
 'use strict';
 
-let libraryInfo = [
-    {
-        file: 'khrang.ogg',
-        author: 'Jerobeam Fenderson',
-        title: 'Khrậng',
-        link: 'https://www.youtube.com/watch?v=vAyCl4IHIz8',
-        swap: true,
-    },
-    {
-        file: 'oscillofun.ogg',
-        author: 'ATOM DELTA',
-        title: 'Oscillofun',
-        link: 'https://www.youtube.com/watch?v=o4YyI6_y6kw',
-        invert: true,
-    },
-    {
-        file: 'alpha_molecule.ogg',
-        author: 'Alexander Taylor',
-        title: 'The Alpha Molecule',
-        link: 'https://www.youtube.com/watch?v=XM8kYRS-cNk',
-        invert: true,
-    },
-];
-
-let libraryDict = {};
-for (let e of libraryInfo) {
-    libraryDict[e.file] = e;
-}
-
-let query = parseq(location.search);
-if (!query.file) {
-    query = libraryInfo[0];
-}
-
-let file = query.file;
-let audioUrl = './woscope-music/' + file;
-let swap = query.swap;
-let invert = query.invert;
 
 let audioCtx = new AudioContext();
+let swap = false;
+let invert = false;
 let audioData = null;
 let quadIndex = null;
 let vertexIndex = null;
@@ -65,12 +29,15 @@ function axhr(url, callback, progress) {
     request.send();
 }
 
-window.onload = function() {
-    let canvas = $('c'),
+module.exports = woscope;
+function woscope(config) {
+    let canvas = config.canvas,
         gl = initGl(canvas),
-        htmlAudio = $('htmlAudio');
+        htmlAudio = config.htmlAudio,
+        audioUrl = config.audioUrl;
 
-    updatePageInfo();
+    swap = config.swap;
+    invert = config.invert;
 
     gl.lineShader = CreateShader(gl, getText('vsLine'), getText('fsLine'));
     gl.blurShader = CreateShader(gl, getText('vsBlurTranspose'), getText('fsBlurTranspose'));
@@ -105,7 +72,6 @@ window.onload = function() {
     };
     progressLoop();
 
-    htmlAudio.volume = 0.5;
     htmlAudio.src = audioUrl;
 
     axhr(audioUrl, function(buffer) {
@@ -121,59 +87,6 @@ window.onload = function() {
         progress = e.total ? e.loaded / e.total : 1.0;
         console.log('progress: ' + e.loaded + ' / ' + e.total);
     });
-};
-
-function parseq(search) {
-    search = search.replace(/^\?/, '');
-    let obj = {};
-    for (let pair of search.split('&')) {
-        pair = pair.split('=');
-        obj[decodeURIComponent(pair[0])] =
-            pair.length > 1 ? decodeURIComponent(pair[1]) : true;
-    }
-    return obj;
-}
-
-function dumpq(obj) {
-    return Object.keys(obj).map(function(key) {
-        if (obj[key] === true) {
-            return encodeURIComponent(key);
-        } else {
-            return encodeURIComponent(key) + '=' + encodeURIComponent(obj[key]);
-        }
-    }).join('&');
-}
-
-function updatePageInfo() {
-    if (file in libraryDict) {
-        let info = libraryDict[file],
-            text = document.createTextNode(info.author + ' — ' + info.title + ' '),
-            songInfo = $('songInfo'),
-            a = document.createElement('a'),
-            linkText = document.createTextNode('[link]');
-
-        a.appendChild(linkText);
-        a.href = info.link;
-        songInfo.innerHTML = '';
-        songInfo.appendChild(text);
-        songInfo.appendChild(a);
-    }
-
-    let ul = $('playList');
-    ul.innerHTML = '';
-    for (let song of libraryInfo) {
-        let a = document.createElement('a'),
-            li = document.createElement('li');
-        a.appendChild(document.createTextNode(song.title));
-
-        let q = {file: song.file};
-        if (song.swap) { q.swap = true; }
-        if (song.invert) { q.invert = true; }
-        a.href = '?' + dumpq(q);
-
-        li.appendChild(a);
-        ul.appendChild(li);
-    }
 }
 
 function initGl(canvas) {
